Testing on non-float64 dtype

+++ heat/core/tests/test_statistics.py
@@ -891,8 +891,8 @@ class TestStatistics(BasicTest):
             ht.minimum(random_volume_1, random_volume_2, out=output)
 
     def test_percentile(self):
-        # test local, distributed, split/axis combination, TODO no data on process
-        x_np = np.arange(10 * 10 * 10, dtype=np.float64).reshape(10, 10, 10)
+        # test local, distributed, split/axis combination, TODO no data on process, Issue #568
+        x_np = np.arange(10 * 10 * 10).reshape(10, 10, 10)
         x_ht = ht.array(x_np, device=ht_device)
         x_ht_split0 = ht.array(x_np, split=0, device=ht_device)
         x_ht_split1 = ht.array(x_np, split=1, device=ht_device)
